+++ packages/dialect-react/components/DialectContext/index.tsx
@@ -18,11 +18,7 @@ import {
   getDialectAddressWithOtherMember,
   sendMessage,
 } from '../../api';
-import {
-  noAccount,
-  ParsedErrorData,
-  ParsedErrorType,
-} from '../../utils/errors';
+import { ParsedErrorData, ParsedErrorType } from '../../utils/errors';
 import {
   connected,
   getMessageHash,
@@ -93,6 +89,7 @@ type DialectContextType = {
   sendMessage: (text: string, encrypted?: boolean) => Promise<void>;
   sendingMessage: boolean;
   sendMessageError: ParsedErrorData | null;
+  checkUnreadMessages: (threadId: string) => boolean;
 };
 
 const DialectContext = createContext<DialectContextType | null>(null);
@@ -125,7 +122,7 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
   const [sendMessageError, setSendMessageError] =
     React.useState<ParsedErrorData | null>(null);
 
-  const { wallet, program, walletName } = useApi();
+  const { wallet, program, walletName, getLastReadMessage } = useApi();
   const isWalletConnected = connected(wallet);
   const [messages, setMessages] = React.useState<Message[]>([]);
 
@@ -429,6 +426,11 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
     [getEncryptionProps, isWalletConnected, program, dialect, mutateDialect]
   );
 
+  const checkUnreadMessages = (threadId: string) => {
+    const lastReadMessage = getLastReadMessage(threadId);
+    return lastReadMessage !== messages[0]?.timestamp && messages.length > 0;
+  };
+
   // const messages = mockMessages;
   const isDialectAvailable = Boolean(dialect);
   const isMetadataAvailable = Boolean(metadata);
@@ -462,6 +464,7 @@ export const DialectProvider = (props: PropsType): JSX.Element => {
     sendMessage: sendMessageWrapper,
     sendingMessage,
     sendMessageError,
+    checkUnreadMessages,
   };
 
   return (