+++ packages/dialect-react-ui/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [UNRELEASED]
 
+- added logic to notfication ui to display new notification badge
 - fix: open thread if user entered address with whom thread already exists
+
 
 ## [0.8.9] - 2022-06-07